+++ IMSWebApi/ViewModel/MaterialQuotation/VMTrnMaterialQuotationItem.cs
@@ -33,7 +33,7 @@ namespace IMSWebApi.ViewModel
         public decimal discountPercentage { get; set; }
         public int amount { get; set; }
         public decimal rateWithGST { get; set; }
-        public decimal amountWithGST { get; set; }
+        public int amountWithGST { get; set; }
         public Nullable<int> gst { get; set; }
         public string status { get; set; }
